Update dependencies-and-umbrella-apps.markdown
Adds a few more examples to for including dependencies via Git and local dependencies; adds a few more links to https://hexdocs.pm/mix/Mix.Tasks.Deps.html.

For future consideration: I think this page would benefit from a more substantial rewrite: the primary focus should be the demonstration of the various ways to include dependencies.  The distinction between internal and external dependencies is more philosophical and the discussion thereof should not interrupt the demonstration of the various syntax flavors.  In other words, the focus should be on the _HOW_; the discussion of the _WHAT_ (i.e. internal vs. external) should be in its own paragraph/section.